refactor: extract Timeline to a separate component
