+++ packages/core/src/components/SpanCard/index.tsx
@@ -8,17 +8,16 @@ import {
   type MouseEvent,
 } from "react";
 
-import type { SpanCardType } from "../types/span";
-import type { ColorVariant } from "../types/ui.ts";
+import type { SpanCardType } from "../../types/span";
 
-import { formatDuration } from "../services/calculate-duration.ts";
 import {
   getSpanCategoryIcon,
   getSpanCategoryLabel,
   getSpanCategoryTheme,
-} from "../utils/ui";
-import { Avatar, type AvatarProps } from "./Avatar";
-import { Badge } from "./Badge";
+} from "../../utils/ui";
+import { Avatar, type AvatarProps } from "../Avatar";
+import { Badge } from "../Badge";
+import { SpanCardTimeline } from "./components/Timeline";
 
 const LAYOUT_CONSTANTS = {
   MARGIN_LEVEL_STEP: 20,
@@ -32,19 +31,6 @@ const STATUS_COLORS = {
   running: "bg-violet-500 dark:bg-violet-700",
   warning: "bg-yellow-500 dark:bg-yellow-700",
 } as const;
-
-const timelineBgColors: Record<ColorVariant, string> = {
-  purple: "bg-purple-400 dark:bg-purple-600",
-  indigo: "bg-indigo-400 dark:bg-indigo-600",
-  orange: "bg-orange-400 dark:bg-orange-600",
-  teal: "bg-teal-400 dark:bg-teal-600",
-  cyan: "bg-cyan-400 dark:bg-cyan-600",
-  sky: "bg-sky-400 dark:bg-sky-600",
-  yellow: "bg-yellow-400 dark:bg-yellow-600",
-  emerald: "bg-emerald-400 dark:bg-emerald-600",
-  red: "bg-red-400 dark:bg-red-600",
-  gray: "bg-gray-400 dark:bg-gray-600",
-};
 
 interface SpanCardProps {
   data: SpanCardType;
@@ -190,7 +176,7 @@ const SpanCardContent: FC<{
 
       <div className="flex items-center justify-start space-x-1">
         <Badge
-          iconStart={<Icon className="h-3 w-3" />}
+          iconStart={<Icon className="size-2.5" />}
           theme={getSpanCategoryTheme(data.type)}
           size="xs"
         >
@@ -208,39 +194,6 @@ const SpanCardContent: FC<{
         </Badge>
       </div>
     </div>
-  );
-};
-
-const SpanCardTimeline: FC<{
-  startTime: Date;
-  endTime: Date;
-  minStart: number;
-  maxEnd: number;
-  theme: ColorVariant;
-}> = ({ startTime, endTime, minStart, maxEnd, theme }) => {
-  const startMs = +startTime;
-  const endMs = +endTime;
-  const totalRange = maxEnd - minStart;
-  const startPercent = ((startMs - minStart) / totalRange) * 100;
-  const widthPercent = ((endMs - startMs) / totalRange) * 100;
-
-  return (
-    <span className="flex w-full items-center">
-      <span className="relative h-3.5 flex-1 rounded bg-gray-100 px-1">
-        <span className="pointer-events-none absolute inset-x-1 top-1/2 h-1.5 -translate-y-1/2">
-          <span
-            className={`absolute h-full rounded-sm ${timelineBgColors[theme]}`}
-            style={{
-              left: `${startPercent}%`,
-              width: `${widthPercent}%`,
-            }}
-          />
-        </span>
-      </span>
-      <span className="ml-2 w-10 whitespace-nowrap text-right text-xs">
-        {formatDuration(endMs - startMs)}
-      </span>
-    </span>
   );
 };
